update randomized encoder

+++ pipelines/matrix/src/matrix/pipelines/embeddings/encoders.py
@@ -96,7 +96,7 @@ class RandomizedEncoder(AttributeEncoder):
         """
         df = df.copy()
         # Generate random embeddings
-        df["embedding"] = [np.random.rand(self._embedding_dim) for _ in range(len(df))]
+        df["embedding"] = [np.random.rand(self._embedding_dim).astype(np.float32) for _ in range(len(df))]
         df = df.drop(columns=["text_to_embed"])
         return df
